Fixed rings
