--- CQ-4K-40K Rings/write_svg.py
@@ -1,6 +1,0 @@
-def write_svg(part, name, direction):
-    with open(f'./svg/{name}','w') as f:
-        f.write(part.toSvg(opts={"projectionDir": direction,
-                              "showHidden": False,
-                              "height": 800,
-                              "width": 800,}))
