+++ outputs/m4K.step
@@ -1,7 +1,7 @@
 ISO-10303-21;
 HEADER;
 FILE_DESCRIPTION(('Open CASCADE Model'),'2;1');
-FILE_NAME('Open CASCADE Shape Model','2021-06-22T15:52:48',('Author'),(
+FILE_NAME('Open CASCADE Shape Model','2021-03-25T07:43:53',('Author'),(
     'Open CASCADE'),'Open CASCADE STEP processor 7.5','Open CASCADE 7.5'
   ,'Unknown');
 FILE_SCHEMA(('AUTOMOTIVE_DESIGN { 1 0 10303 214 1 1 1 1 }'));
@@ -15,8 +15,8 @@ DATA;
 #4 = PRODUCT_DEFINITION_SHAPE('','',#5);
 #5 = PRODUCT_DEFINITION('design','',#6,#9);
 #6 = PRODUCT_DEFINITION_FORMATION('','',#7);
-#7 = PRODUCT('Open CASCADE STEP translator 7.5 3',
-  'Open CASCADE STEP translator 7.5 3','',(#8));
+#7 = PRODUCT('Open CASCADE STEP translator 7.5 22',
+  'Open CASCADE STEP translator 7.5 22','',(#8));
 #8 = PRODUCT_CONTEXT('',#2,'mechanical');
 #9 = PRODUCT_DEFINITION_CONTEXT('part definition',#2,'design');
 #10 = ADVANCED_BREP_SHAPE_REPRESENTATION('',(#11,#15),#3077);
